Enable PHP7 on Ubuntu 14.04
PHP7 come with Ubuntu 16.04 but since we can't run it yet we can use
this hack.

I used this guide as a help to install PHP7.
https://www.digitalocean.com/community/tutorials/how-to-upgrade-to-php-7-on-ubuntu-14-04

All WP unit tests succeed and suite run in 60sec instead of 120sec.
50%  performance gain, and we can code with new PHP7 features :)

Puppet script is really dirty but works.
We'll be able to ditch it once we get native Ubuntu 16.04.